+++ agent/connection_helper.go
