+++ agent/topology.go
@@ -1,4 +1,4 @@
-package network
+package agent
 
 import (
 	"fmt"
